requirements.txt
requests
beautifulsoup4
python-dotenv
python-telegram-bot
selenium
webdriver-manager